Fixed AmbisonicPanner plugin so it includes Encoder/Decoder as source not linked lib
There were various bugs introduced by linking to the Encoder and Decoder plugins as a
library.  This is not the way JUCE intends for things to work.

As a result, I changed  CMakeLists to include all the source files from the encoder and
decoder plugins into the Panner plugin rather then linking statically

Also made various bug fixes:

- Fixed ambiguous template deduction in PanModulator
- Made various float casts explicit to suppress loss of precision warnings
- Fixed bug in LFOSingleController where an unsupported '\s' escape sequence was used
  instead of a space character
- Fixed bugs in various editor components where components would add themselves as
  a listener, but weren't removing themselves upon destruction.  This created errors when
  an editor would be closed, destroyed and then reconstructed.
- Changed the AmbisonicPanner to use a PlugCreator library
  (see PlugFilterCreator/CMakeLists.txt)

+++ Panner/Source/GUI/AutoPanComponent.h
@@ -31,7 +31,7 @@ class AutoPanComponent : public juce::Component, public juce::ChangeBroadcaster
 
     float getRateSetting()
     {
-      return quickRateKnob->getValue();
+      return (float)quickRateKnob->getValue();
     }
 
     void setRateSetting(float rate, bool sendNotificaiton)
